pyproject.toml
[project]
name = "mfcl"
version = "0.1.0"
description = "Minimal, modular self-supervised learning components"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "torch>=2.2",
  "torchvision>=0.17",
  "timm>=0.9",
  "hydra-core>=1.3",
  "omegaconf>=2.3",
  "numpy>=1.23",
  "scipy>=1.10",
  "scikit-learn>=1.2",
  "matplotlib>=3.6",
]

[project.optional-dependencies]
test = [
  "pytest>=7.0",
  "hypothesis>=6.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["mfcl*"]

